Couple more improvements
