+++ expression-builder/src/Operator.tsx
@@ -9,22 +9,18 @@ import {
   OperatorParameterMetadata,
   EvaluatorNode,
   FragmentParameterMetadata,
-  Operator as OpType,
+
   // json-edit-react
   CustomNodeProps,
-  IconEdit,
   IconOk,
   IconCancel,
 } from './_imports'
+import { DisplayBar } from './DisplayBar'
 import { Select, SelectOption } from './Select'
-import { Icons } from './Icons'
-import { getButtonFontSize, getCurrentOperator, getDefaultValue } from './helpers'
+import { getCurrentOperator, getDefaultValue } from './helpers'
 import { NodeTypeSelector } from './NodeTypeSelector'
 import { useCommon } from './useCommon'
 import { cleanOperatorNode, getAvailableProperties } from './validator'
-import { operatorDisplay } from './operatorDisplay'
-
-const README_URL = 'https://github.com/CarlosNZ/fig-tree-evaluator?tab=readme-ov-file#'
 
 export interface OperatorProps {
   figTree: FigTreeEvaluator
@@ -104,6 +100,7 @@ export const Operator: React.FC<CustomNodeProps<OperatorProps>> = (props) => {
       ) : (
         <DisplayBar
           name={thisOperator}
+          description={operatorData.description}
           setIsEditing={() => setIsEditing(true)}
           evaluate={evaluate}
           isLoading={loading}
@@ -125,98 +122,6 @@ export const Operator: React.FC<CustomNodeProps<OperatorProps>> = (props) => {
             onEdit(newNode, expressionPath)
           }}
         />
-      )}
-    </div>
-  )
-}
-
-interface DisplayBarProps {
-  name: OperatorAlias
-  setIsEditing: () => void
-  evaluate: () => void
-  isLoading: boolean
-  canonicalName: OpType | 'FRAGMENT'
-}
-
-export const DisplayBar: React.FC<DisplayBarProps> = ({
-  name,
-  setIsEditing,
-  evaluate,
-  isLoading,
-  canonicalName = 'CUSTOM_FUNCTIONS',
-}) => {
-  const { backgroundColor, textColor, displayName } = operatorDisplay[canonicalName]
-  const isShorthand = name.startsWith('$')
-  const link = README_URL + canonicalName.toLowerCase() + (canonicalName === 'FRAGMENT' ? 's' : '')
-
-  return (
-    <div className="ft-display-bar">
-      <div className="ft-button-and-edit">
-        <EvaluateButton
-          name={name}
-          backgroundColor={backgroundColor}
-          textColor={textColor}
-          evaluate={evaluate}
-          isLoading={isLoading}
-        />
-        {!isShorthand && (
-          <span onClick={() => setIsEditing()} className="ft-clickable ft-edit-icon">
-            <IconEdit size="1.5em" style={{ color: 'rgb(42, 161, 152)' }} />
-          </span>
-        )}
-      </div>
-      <div className="ft-display-name">
-        <a href={link} target="_blank">
-          {displayName}
-        </a>
-      </div>
-    </div>
-  )
-}
-
-export interface EvaluateButtonProps {
-  name?: string
-  backgroundColor: string
-  textColor: string
-  evaluate: () => void
-  isLoading: boolean
-}
-
-export const EvaluateButton: React.FC<EvaluateButtonProps> = ({
-  name,
-  backgroundColor,
-  textColor,
-  evaluate,
-  isLoading,
-}) => {
-  return (
-    <div
-      className="ft-display-button"
-      style={{ backgroundColor, color: textColor }}
-      onClick={evaluate}
-    >
-      {!isLoading ? (
-        <>
-          {name && (
-            <span
-              className="ft-operator-alias"
-              style={{
-                fontSize: getButtonFontSize(name),
-                fontStyle: 'inherit',
-              }}
-            >
-              {name}
-            </span>
-          )}
-          {Icons.evaluate}
-        </>
-      ) : (
-        <div style={{ width: '100%', textAlign: 'center' }}>
-          <span
-            className="loader"
-            style={{ width: '1.5em', height: '1.5em', borderTopColor: textColor }}
-          ></span>
-        </div>
       )}
     </div>
   )
